+++ testdata/de/dm/intellij/liferay/language/jsp/LiferayTaglibJavascriptFileReferenceContributorTest/js/MyModule.js
@@ -1,0 +1,4 @@
+const MyModule = () => {
+}
+
+export default MyModule;